feat: shorter dashboard link url
